requirements.txt
Flask~=3.1.1
SQLAlchemy~=2.0.41
requests~=2.32.4
python-dotenv~=1.1.1